Script to concatenate fastqs safely
Concatenate_fastqs.py is meant to aid the concatenate_lanes function.
The goal is to concatenate fastq files with headers specifying the
sample, cluster name and library where it came from, and add a unique
identifier (readnum) within the files to be concatenated.
This with the purpose of remove any chance of header duplication when
later concatenating fastq files of a same cluster coming from different
samples (merge_clusters function).
The function takes a list of files, sample id, cluster name, and a list
of the libraries (meant to be one per given fastq file). It opens each
of the files and modifies the header using SeqIO record.id. This
adds the desired string at the beginning of the header title.