Mapping with read number (line number) instead of read ID
STAR has a parameter to use the line number of the read from the BAM file instead of the read ID (so that there are no conflicts with different reads having the same header).
I've changed it now and returned concatenate_lanes to a (better) cat instead of the script - since it's way faster.
Although, it still checks the library type so that's better than before.