main: Fix normalization using only clusters from a specific group
There was a bug where if you had sample names "sample1" and "sample11" (but were only interested in sample1) the program would also read sample11 but wouldn't be able to find it in cluster_sizes, so it would finish unsuccesfully. This was due to a bad regex where i was only checking the sample/group name, but not the name of the cluster itself. I fixed it by subsetting files only if the file name starts the same way as the cluster is called.