Change of the cat for subprocess.call in case slurm == None
subprocess.call doesnt accept the full command of cat infile1 infile2 > outfile. Therefore, I added an if checking if we are running locally or we are submitting to sbatch. If we are not using sbatch, then I used subprocess.call with stdout = outfile. This will affect the assesment of the exitcode in concatenate_lanes at experiment.py