Changes on the testing file structure
Expected results can't be at the same directory of the output, as trusTEr relies on certain file structure.

We used to have the expected results in each of the steps' directories which resulted on the concatenation of the mock files in the merge_clusters step (which was giving files of double length).

I also changed some of the expected results mock files since some of them had typos or were missing a newline at the end of the file (needed as cat ends in new line).

Bamtofastq files were missing a whitespace between the two headers of the reads.